requirements.txt
django-environ==0.11.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
django-storages==1.14.4
django-ckeditor==6.7.1
djangorestframework-simplejwt
